Had to adjust __new__ to return the creation of an  object when being inherited by another class. The previous behavior returned the  everytime we inherited from AbstractFlow, but now I ensured that if the name of the class being inherited is , then AND only then do we throw the error
